chore: move type declarations for risk to index.d.ts
