+++ src/backend/riskManager/RiskExplanation.ts
@@ -1,23 +1,6 @@
 import { Expose } from "class-transformer";
+import type { Explanation, RetentionPeriod, RoleExplanation } from ".";
 import type { Role } from "../Role";
-
-/**
- * A RoleExplanation is a Record with the Role as key and an explanation
- */
-export type RoleExplanation = Record<Role, Explanation>;
-
-/**
- * An Explanation has an translation key and optionally a source string (e.g. another website)
- */
-export type Explanation = {
-  translationKey: string;
-  source?: string;
-};
-
-/**
- * A RetentionPeriod can be either a string (translation key) or a number (retention period in number of days)
- */
-export type RetentionPeriod = string | number;
 
 /**
  *  The RiskExplanation Class that contains the explanations for each risk of a data type (e.g. User Forename)
